VC_Analyst/requirements.txt
google-adk[a2a]>=1.9.0,<2.0.0
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.40.0
requests>=2.31.0
pydantic>=2.6.0
# Agent runtime deps
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
joblib>=1.3.0
serpapi>=0.1.5
# Founder/NN model
tensorflow>=2.14.0
# Tests
pytest>=8.0.0
pytest-mock>=3.12.0
